Add files via upload
This Jupyter notebook is graphing out the amortization rate for all clothing classes (top, dress, sweater, etc.) to see which rental class of clothing is the most popular among customers. The more time the clothing class is rented out, the faster that class will break even with the purchasing cost of that class (faster amortization rate).
For example, if a class is more popular, the amortization rate for that class will be faster, meaning the line of that class on the graph will have a steeper downward slope.